relocate the tests and stdoutWriter factory to the correct place
